Maybe if I refactor this probject enough times I'll hit some kind of world record.
