+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "hemlock"]
 	path = hemlock
-	url = https://github.com/trashbyte/hemlock
+	url = https://github.com/NotGyro/hemlock
