+++ gestalt/Cargo.toml
@@ -9,27 +9,31 @@ hemlock = { path = "../hemlock" }
 
 anyhow = "1.0.32"
 arr_macro = "0.1.3"
+base16 = "0.2"
 bincode = "1.3.1"
 byteorder = "1.3.4"
 cgmath = "0.17.0"
 chrono = "0.4.15"
-crossbeam-channel = "0.4.3"
+clap = "2.33.3"
+crossbeam-channel = "0.3"
 custom_error = "1.7.1"
 enum_dispatch = "0.3.2"
 glium = "0.27.0"
 glutin = "0.24.1"
 hashbrown = "0.8.2"
 image = "0.23.8"
+laminar = "0.3"
 lazy_static = "1.4.0"
+legion = "0.3.0"
 log = { version = "0.4.11", features = ["std"] }
-nalgebra = "0.21.1"
 num = "0.3.0"
 parking_lot = "0.11.0"
 rand = "0.7"
-ron = "0.6.0"
+ron = "0.6.1"
 rusty_v8 = "0.8.1"
 semver = "0.10.0"
 serde = { version = "1.0.115", features = ["derive"] }
+sodiumoxide = "0.2"
 ustr = { version = "0.7.0", features = ["serialization"] }
 uuid = { version = "0.8.1", features = ["serde", "v4"] }
